+++ build.sh
@@ -1,6 +1,3 @@
 #!/bin/bash
 
-# Build the binary outside src
-cd src
-go build -ldflags "-s -w" -o ..
-cd ..
+go build -ldflags "-s -w"