+++ util/cache.go
@@ -1,9 +1,9 @@
-package utils
+package util
 
 import (
 	"time"
 
-	"github.com/LockBlock-dev/MinePot/typings"
+	"github.com/LockBlock-dev/MinePot/types"
 	"github.com/muesli/cache2go"
 )
 
@@ -17,22 +17,22 @@ func AddToCache(key interface{}, lifeSpan time.Duration, data interface{}) bool 
 	return false
 }
 
-func shouldReport(host string, cooldown int, reportType bool ) bool {
-    item, err := cache2go.Cache("MinePot").Value(host)
+func shouldReport(host string, cooldown int, reportType bool) bool {
+	item, err := cache2go.Cache("MinePot").Value(host)
 	if err != nil {
 		return true
 	}
 
 	if reportType {
 		// Check if the report was reported to AIPDB and if it's older than the cooldown
-		report := item.Data().(typings.Report)
-		if !report.ReportedAIPDB && time.Since(report.Datetime) > (time.Duration(cooldown) * time.Hour) {
+		report := item.Data().(types.Report)
+		if !report.ReportedAIPDB && time.Since(report.Datetime) > (time.Duration(cooldown)*time.Hour) {
 			return true
 		}
 	} else {
 		// Check if the report was reported to the webhook and if it's older than the cooldown
-		report := item.Data().(typings.Report)
-		if !report.ReportedWebhook && time.Since(report.Datetime) > (time.Duration(cooldown) * time.Hour) {
+		report := item.Data().(types.Report)
+		if !report.ReportedWebhook && time.Since(report.Datetime) > (time.Duration(cooldown)*time.Hour) {
 			return true
 		}
 	}