+++ util/logic.go
@@ -1,4 +1,4 @@
-package utils
+package util
 
 import (
 	"fmt"
@@ -8,28 +8,26 @@ import (
 	"os"
 	"time"
 
-	"github.com/LockBlock-dev/MinePot/typings"
+	"github.com/LockBlock-dev/MinePot/types"
 )
 
-
-
-func HandleReport(conn typings.ConnWrapper, addr string) {
+func HandleReport(conn types.ConnWrapper, addr string) {
 	if conn.PacketsReceived >= conn.Config.ReportThreshold {
 		host, _, err := net.SplitHostPort(addr)
 		if err != nil {
-			log.Println(addr + " - Failed to read host from address:", err)
+			log.Println(addr+" - Failed to read host from address:", err)
 		}
 
 		var reportedAIPDB = false
 		if conn.Config.AbuseIPDBReport && shouldReport(host, conn.Config.AbuseIPDBCooldownH, true) {
 			respCode, err := Report(host, conn.Config.AbuseIPDBKey, conn.Config.Port)
 			if err != nil {
-				log.Println(addr + " - Failed to report on AbuseIPDB:", err)
+				log.Println(addr+" - Failed to report on AbuseIPDB:", err)
 			} else if respCode == 200 {
 				reportedAIPDB = true
 			}
 		}
-	
+
 		var reportedWebhook = false
 		if conn.Config.WebhookReport && shouldReport(host, conn.Config.WebhookCooldownH, false) {
 			err := SendWebhook(
@@ -40,7 +38,7 @@ func HandleReport(conn typings.ConnWrapper, addr string) {
 				conn.DidPing,
 			)
 			if err != nil {
-				log.Println(addr + " - Failed to report on webhook:", err)
+				log.Println(addr+" - Failed to report on webhook:", err)
 			} else {
 				reportedWebhook = true
 			}
@@ -49,27 +47,27 @@ func HandleReport(conn typings.ConnWrapper, addr string) {
 		AddToCache(
 			host,
 			// The maximum time between the AbuseIPDB and Webhook report (in hours)
-			time.Duration(math.Max(float64(conn.Config.AbuseIPDBCooldownH), float64(conn.Config.WebhookCooldownH))) * time.Hour,
-			typings.Report{
-				Datetime: time.Now(),
-				PacketsCount: conn.PacketsReceived,
-				ReportedAIPDB: reportedAIPDB,
+			time.Duration(math.Max(float64(conn.Config.AbuseIPDBCooldownH), float64(conn.Config.WebhookCooldownH)))*time.Hour,
+			types.Report{
+				Datetime:        time.Now(),
+				PacketsCount:    conn.PacketsReceived,
+				ReportedAIPDB:   reportedAIPDB,
 				ReportedWebhook: reportedWebhook,
-				Handshake: conn.DidHandshake,
-				Ping: conn.DidPing,
+				Handshake:       conn.DidHandshake,
+				Ping:            conn.DidPing,
 			},
 		)
-	
+
 		if conn.Config.WriteHistory {
 			// Open history file
 			file, err := os.OpenFile(conn.Config.HistoryFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644) // 644 = rw-,r--,r--
 			if err != nil {
-				log.Println(addr + " - Failed to open history file:", err)
+				log.Println(addr+" - Failed to open history file:", err)
 			}
 			defer file.Close()
-	
+
 			t := time.Now()
-	
+
 			_, err = file.WriteString(fmt.Sprintf(
 				"%s,%s,%d,%t,%t,%t\n",
 				t.Format("2006-01-02 15:04:05"),
@@ -80,7 +78,7 @@ func HandleReport(conn typings.ConnWrapper, addr string) {
 				conn.DidPing,
 			))
 			if err != nil {
-				log.Println(addr + " - Failed to write history:", err)
+				log.Println(addr+" - Failed to write history:", err)
 			}
 		}
 	}