+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2024-07-17
+
+### Added
+
+-   Artificial random ping
+-   Random protocol/version
+-   Version name with client protocol mirroring
+
+### Changed
+
+-   Project layout
+-   Minecraft protocol implementation from `go-mc`
+-   Added a build step to get a smaller Docker image
+
 ## [2.0.0] - 2023-07-20
 
 ### Changed